database seeders

+++ database/migrations/2021_09_20_164258_create_vehicles_table.php
@@ -18,7 +18,7 @@ class CreateVehiclesTable extends Migration
             $table->timestamps();
             $table->string('name');
             $table->string('registration_number');
-            $table->decimal('mileage');
+            $table->integer('mileage');
         });
     }
